requirements.txt
pyTelegramBotAPI==4.12.0
psycopg2-binary==2.9.6